feat: 제품 및 창고 코드 조회 기능 추가 및 가격 타입 수정

+++ src/main/java/com/example/kafkaorder/entity/Product.java
@@ -29,7 +29,7 @@ public class Product {
     private String name;
 
     @Column(nullable = false)
-    private BigDecimal price;
+    private int price;
 
     @Lob
     private String description;
